# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: NextEvent Go API
  description: |
    Comprehensive API for the NextEvent platform, providing event management, 
    content management, news publishing, and video streaming capabilities.
    
    This API maintains full compatibility with the existing .NET implementation
    while providing enhanced performance and scalability.
  version: 2.0.0
  contact:
    name: NextEvent Development Team
    email: dev@nextevent.com
  license:
    name: Proprietary
    url: https://nextevent.com/license
servers:
  - url: https://api.nextevent.com/v2
    description: Production server
  - url: https://staging-api.nextevent.com/v2
    description: Staging server
  - url: http://localhost:8080/api/v2
    description: Development server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /articles:
    get:
      tags:
        - Articles
      summary: List articles
      description: Retrieve a paginated list of articles with filtering options
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: search
          in: query
          description: Search term for article titles or content
          schema:
            type: string
        - name: status
          in: query
          description: Filter by article status
          schema:
            $ref: '#/components/schemas/ArticleStatus'
        - name: authorId
          in: query
          description: Filter by author ID
          schema:
            type: string
            format: uuid
        - name: categoryId
          in: query
          description: Filter by category ID
          schema:
            type: string
            format: uuid
        - name: includeContent
          in: query
          description: Include full article content in response
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of articles retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Articles
      summary: Create article
      description: Create a new article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleCreateRequest'
      responses:
        '201':
          description: Article created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # News Management API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    Page:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "123e4567-e89b-12d3-a456-426614174000"
    Timestamp:
      type: string
      format: date-time
      example: "2023-12-01T10:30:00Z"

    # Image Management Schemas
...
    ArticleStatus:
      type: string
      enum: [draft, published, archived, deleted]
      description: Status of article
    ArticleType:
      type: string
      enum: [standard, featured, news, blog, guide]
      description: Type of article
    ArticleResponse:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
          example: "Getting Started with NextEvent"
        subtitle:
          type: string
          example: "A comprehensive guide to event management"
        summary:
          type: string
          example: "Learn how to create and manage events effectively"
        content:
          type: string
          example: "# Introduction\n\nThis guide will help you..."
        type:
          $ref: '#/components/schemas/ArticleType'
        status:
          $ref: '#/components/schemas/ArticleStatus'
        slug:
          type: string
          example: "getting-started-with-nextevent"
        featuredImageId:
          $ref: '#/components/schemas/UUID'
        thumbnailId:
          $ref: '#/components/schemas/UUID'
        authorId:
          $ref: '#/components/schemas/UUID'
        editorId:
          $ref: '#/components/schemas/UUID'
        categoryId:
          $ref: '#/components/schemas/UUID'
        tags:
          type: array
          items:
            type: string
          example: ["guide", "tutorial", "events"]
        metaTitle:
          type: string
        metaDescription:
          type: string
        keywords:
          type: string
        readTime:
          type: integer
          description: Estimated read time in minutes
        viewCount:
          type: integer
          format: int64
        likeCount:
          type: integer
          format: int64
        shareCount:
          type: integer
          format: int64
        publishedAt:
          $ref: '#/components/schemas/Timestamp'
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
    ArticleCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        subtitle:
          type: string
          maxLength: 1000
        summary:
          type: string
          maxLength: 2000
        content:
          type: string
        type:
          $ref: '#/components/schemas/ArticleType'
        slug:
          type: string
          maxLength: 500
        featuredImageId:
          $ref: '#/components/schemas/UUID'
        thumbnailId:
          $ref: '#/components/schemas/UUID'
        categoryId:
          $ref: '#/components/schemas/UUID'
        tags:
          type: array
          items:
            type: string
        metaTitle:
          type: string
          maxLength: 500
        metaDescription:
          type: string
          maxLength: 1000
        keywords:
          type: string
          maxLength: 1000
        scheduledAt:
          $ref: '#/components/schemas/Timestamp'
      required:
        - title
        - content
        - type
    ArticleListResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/ArticleResponse'
        total:
          type: integer
          format: int64
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer

    # News Management Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string
            timestamp:
              $ref: '#/components/schemas/Timestamp'
            requestId:
              type: string
              format: uuid
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Validation failed"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: "title"
                  message:
                    type: string
                    example: "Title is required"
                  code:
                    type: string
                    example: "REQUIRED"
            timestamp:
              $ref: '#/components/schemas/Timestamp'
            requestId:
              type: string
              format: uuid
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
              timestamp: "2023-12-01T10:30:00Z"
              requestId: "123e4567-e89b-12d3-a456-426614174000"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "INTERNAL_ERROR"
              message: "An unexpected error occurred"
              timestamp: "2023-12-01T10:30:00Z"
              requestId: "123e4567-e89b-12d3-a456-426614174000"
...